feat: serializer finished, test still missing

+++ nbt_lib/src/writer/java.rs
@@ -29,7 +29,7 @@ impl NbtWrite for Java {
     fn write_compound(writer: &mut Vec<u8>, name: Option<&String>, data: &[(String, NbtValue)]) -> NbtResult<()> {
         Self::write_nbt_string(writer, name.unwrap_or(&"".to_string()));
         for (key, value) in data {
-            writer.push(value.tag());
+            writer.push(value.tag() as u8);
             Self::write_nbt_string(writer, key);
             match value {
                 NbtValue::Byte(v) => writer.push(*v as u8),
@@ -56,17 +56,17 @@ impl NbtWrite for Java {
     fn write_to(value: &crate::NbtValue, buff: &mut Vec<u8>) -> crate::error::NbtResult<()> {
         match value {
             NbtValue::Compound(name, data) => {
-                buff.push(value.tag());
+                buff.push(value.tag() as u8);
                 Self::write_compound(buff, name.as_ref(), data)?
             }
-            v => return Err(NbtError::WrongRootType(v.tag())),
+            v => return Err(NbtError::WrongRootType(v.tag() as u8)),
         }
         Ok(())
     }
 
     #[cfg_attr(feature = "inline_read", inline)]
     fn write_to_with_name(name: &str, value: &crate::NbtValue, buff: &mut Vec<u8>) -> crate::error::NbtResult<()> {
-        buff.push(value.tag());
+        buff.push(value.tag() as u8);
         Self::write_nbt_string(buff, name);
         Self::write_to(value, buff)
     }
@@ -81,7 +81,7 @@ impl NbtWrite for Java {
         if !data.iter().all(|x| x.tag() == tag) {
             return Err(NbtError::ListTypeNotSame(data.iter().map(|x| x.tag()).collect()));
         }
-        writer.push(tag);
+        writer.push(tag as u8);
         writer.extend_from_slice(&(data.len() as i32).to_be_bytes());
 
         for d in data {
@@ -110,7 +110,7 @@ impl NbtWrite for Java {
         match value {
             NbtValue::String(str) => Ok(Java::write_nbt_string(writer, str)),
             NbtValue::Compound(_, _) => Java::write_to(value, writer),
-            v => Err(NbtError::WrongRootType(v.tag())),
+            v => Err(NbtError::WrongRootType(v.tag() as u8)),
         }
     }
 }
